requirements.txt
django
django-crispy-forms
django-bootstrap4
django-filter
django-bootstrap-form
unidecode
django-admin-interface
django-jet
google-api-python-client
django-admin-autocomplete-filter
django-tinymce